Code consolidated for MiKo_1040, MiKo_1041 and MiKo_1070
